rename the util class and added into new util package
