+++ src/main/java/com/identity/service/service/RoleService.java
@@ -1,10 +1,11 @@
 package com.identity.service.service;
 
 
-import com.identity.service.EsUtil;
+import com.identity.service.util.Util;
 import com.identity.service.models.Roles;
 import com.identity.service.repository.RoleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,14 +18,14 @@ public class RoleService {
 
 	public List<Roles> findAll() {
 		Roles roles = new Roles();
-		roles.setFlagId(EsUtil.ROLE_FIND_ALL_FLAG);
+		roles.setFlagId(Util.ROLE_FIND_ALL_FLAG);
 		return roleDaoService.procAddRole(roles);
 	}
 
 	public Roles findOne(short id) {
 
 		Roles roles = new Roles();
-		roles.setFlagId(EsUtil.ROLE_FIND_ONE_FLAG);
+		roles.setFlagId(Util.ROLE_FIND_ONE_FLAG);
 		roles.setRoleId(id);
 		List<Roles> roleList = roleDaoService.procAddRole(roles);
 		if (roleList.size() > 0)
@@ -35,7 +36,7 @@ public class RoleService {
 	}
 
 	public Roles save(Roles role) {
-		role.setFlagId(EsUtil.ROLE_SAVE_FLAG);
+		role.setFlagId(Util.ROLE_SAVE_FLAG);
 
 		List<Roles> roleList = roleDaoService.procAddRole(role);
 		if (roleList.size() > 0)
@@ -47,7 +48,7 @@ public class RoleService {
 	}
 
 	public Roles update(Roles role) {
-		role.setFlagId(EsUtil.ROLE_UPDATE_FLAG);
+		role.setFlagId(Util.ROLE_UPDATE_FLAG);
 
 		List<Roles> roleList = roleDaoService.procAddRole(role);
 		if (roleList.size() > 0)
