+++ src/main/java/com/identity/service/util/Util.java
@@ -1,6 +1,6 @@
-package com.identity.service;
+package com.identity.service.util;
 
-public class EsUtil {
+public class Util {
 
     public final static short ROLE_FIND_ALL_FLAG = 1;
     public final static short ROLE_FIND_ONE_FLAG = 1;
